Migrate parser and serializer FieldType methods into subpackage
