+++ supyr_struct/field_types.py
@@ -31,6 +31,7 @@ from supyr_struct.buffer import BytesBuffer, BytearrayBuffer
 from supyr_struct import blocks
 from supyr_struct.defs.constants import NAME, SIZE, TYPE, ENTRIES,\
      byteorder_char
+from supyr_struct.defs.sanitizers import *
 from supyr_struct.defs.frozen_dict import FrozenDict
 
 # ######################################
